+++ examples/test-site/src/components/Filter/index.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright © 2020 Johnson & Johnson
+ * Copyright © 2019 Johnson & Johnson
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,68 +12,15 @@
  * limitations under the License.
  */
 
-import React, { useState } from 'react';
 import { asTaggableItem, withFilterByTags } from '@bodiless/components';
-import {
-  H2, Span, Button, addClasses,
-} from '@bodiless/fclasses';
+import { Span, addClasses, Button } from '@bodiless/fclasses';
 import { flow } from 'lodash';
-import { TagType } from '@bodiless/core';
 
-const TagButton = addClasses('px-2 mb-2 mr-2 border border-gray-600')(Button);
 const TagSpan = addClasses('px-2 mb-2 mr-2')(Span);
-const getSuggestions = () => [
-  { id: 0, name: 'foo' },
-  { id: 1, name: 'baz' },
-  { id: 2, name: 'bat' },
-  { id: 3, name: 'bar' },
-];
+const TagButton = addClasses('px-2 mb-2 mr-2 border border-gray-600')(Button);
 const TaggableFilterableItem = flow(
   withFilterByTags,
   asTaggableItem(),
 )(TagSpan);
-const TaggableFilterSelector = () => {
-  const [tags, setTags] = useState<TagType[]>([]);
-  const Buttons = getSuggestions().map(tag => (
-    <TagButton key={tag.id} onClick={() => setTags([tag])}>
-      {tag.name}
-    </TagButton>
-  ));
-  const props = {
-    getSuggestions,
-    placeholder: 'Add or create',
-    formTitle: 'Groups',
-    seeAllText: 'See all groups',
-    formBodyText: 'Select from available tags:',
-    allowNew: true,
-    noSuggestionsText: 'No suggestions found',
-  };
-  const FilterableItemsList = getSuggestions().map(tag => (
-    <TaggableFilterableItem
-      id={tag.name}
-      {...props}
-      nodeKey={tag.name}
-      key={tag.id}
-      selectedTags={tags}
-    >
-      {tag.name}
-    </TaggableFilterableItem>
-  ));
-  return (
-    <div>
-      <div>
-        <H2>Select a tag to filter by</H2>
-        {Buttons}
-        <TagButton id="show-all" onClick={() => setTags([])}>
-          ALL
-        </TagButton>
-      </div>
-      <div>
-        <h2>Filtered Components</h2>
-        {FilterableItemsList}
-      </div>
-    </div>
-  );
-};
-
-export default TaggableFilterSelector;
+export { TagButton };
+export default TaggableFilterableItem;
