+++ packages/bodiless-components/src/index.ts
@@ -23,7 +23,7 @@ import {
   TitleProps as ListTitleProps,
   ListDesignableComponents,
 } from './List/types';
-import { asTaggableItem } from './Taggable/Item';
+import asTaggableItem from './Taggable/asTaggableItem';
 import withListTitle from './List/withListTitle';
 import asEditableList from './List/asEditableList';
 import asBasicSublist from './List/asBasicSublist';
